\name{SVILD-package}
\alias{SVILD-package}

\docType{package}
\title{Summary Variability of Irregularly Longitudinal Data}
\description{
	This package summarize glucose variability of irregularly surgical longitudinal
	data and includes functions to create important indices of glucose variability such as
	time weighted average, percentage of time within range, and etc.
}
\details{
	\tabular{ll}{
		Package: \tab SVILD\cr
		Type: \tab Package\cr
		Version: \tab 1.0\cr
		Date: \tab 2010-04-28\cr
		License: \tab GPL-2\cr
		LazyLoad: \tab yes\cr
}

Index:

print		print the object
summary		summary of basic statistics of an object
plot		randomly plot individuals with lowess functions
twa		calculate time weighted average using different interpolate methods
pcttime		calculate perecent time within ranges
hgi		calculate Hyperglycaemic Index 
pgi		calculate Glycemic penalty index 
gri		calculate Glycemic risk index
}

\author{
Dongsheng Yang, Anne S. Tang
Maintained by: Dongsheng Yang <yangsky66@yahoo.com>
}
\references{

	Mathijs Vogelzang etc 2004. Hyperglycaemic index as a tool to assess glucose control: a retrospective study
	Tom Van Herpe etc 2008. Glycemic penalty index for adequately assessing and comparing different blood glucose control algorithms
	Kovatchev BP etc 2006. Evaluation of a New Measure of Blood Glucose Variability in Diabetes
}

\keyword{ package }



