\name{SVILD-package}
\alias{SVILD-package}

\docType{package}
\title{Summaries of Variability of Irregularly Spaced Longitudinal Data}
\description{
	This package summarize glucose variability of irregularly spaced surgical longitudinal
	data and includes functions to create important indices of glucose variability such as
	time weighted average, percentage of time within range, and etc.
}
\details{
	\tabular{ll}{
	Package: \tab SVILD\cr
	Type: \tab Package\cr
	Version: \tab 1.0\cr
	Date: \tab 2010-04-28\cr
	License: \tab GPL-2\cr
	LazyLoad: \tab yes\cr
	}

Index:

	\tabular{ll}{

	print \tab print the object\cr
	summary \tab summary of basic statistics of an object\cr
	plot \tab randomly plot individuals with lowess functions\cr
	twa  \tab calculate time weighted average using different interpolate methods\cr
	pcttime  \tab calculate perecent time within ranges\cr
	hgi  \tab calculate Hyperglycaemic Index\cr 
	gpi \tab calculate Glycemic penalty index\cr 
	gri  \tab calculate Glycemic risk index\cr

	}
}


\author{
Dongsheng Yang, Anne S. Tang.
Maintained by: Dongsheng Yang <yangsky66@yahoo.com>
}
\references{

	Mathijs Vogelzang, et al. Hyperglycaemic index as a tool to assess glucose control: a retrospective study. \emph{Crit Care} Jun; (8)3:151-2. 

	Tom Van Herpe, et al. Glycemic penalty index for adequately assessing and comparing different blood glucose control algorithms.  \emph{Crit Care} 2008; (12)2:133.

	Kovatchev BP, et al. Evaluation of a New Measure of Blood Glucose Variability in Diabetes. \emph{Diabetes Care} 2006 Nov; (29)11:2433-8.
}

\keyword{ package }



